Moved Notification to avoid annoyances with Foundation
