+++ Sources/PostgreSQL/Connection.swift
@@ -155,6 +155,24 @@ public final class Connection: ConnInfoInitializable {
     
     // MARK: - LISTEN/NOTIFY
     
+    public struct Notification {
+        public let pid: Int
+        public let channel: String
+        public let payload: String?
+        
+        init(pgNotify: PGnotify) {
+            channel = String(cString: pgNotify.relname)
+            pid = Int(pgNotify.be_pid)
+            
+            if pgNotify.extra != nil {
+                payload = String(cString: pgNotify.extra)
+            }
+            else {
+                payload = nil
+            }
+        }
+    }
+    
     /// Registers as a listener on a specific notification channel.
     ///
     /// - Parameters:
